Add content SHA1 to file info if provided
B2 does not provide the contentSha1 value for large files, instead returning
"none". This can cause confusion for many users [1], and the fact that the
B2 CLI does not populate the large_file_sha1 file info entry is arguably a
bug.

This implementation only adds the large_file_sha1 file info entry if it is
not already present, and the content SHA1 is known in advance (i.e. it will
not cause a SHA1 to be computed if it is not provided). This works with B2
CLI when specifying the "--sha1" option, which without this, would silently
swallow the provided SHA1, which the user would likely not notice until
_after_ the potentially lengthy upload was complete.

[1] https://github.com/Backblaze/B2_Command_Line_Tool/issues/604