Additional tests for listing files / versions
The test checks whether empty response with nextFileName is properly
handled by the SDK. We provide a bucket implementation that ensures
that every first request for a given file name returns no files
and only points to the same file again. Using that we're performing
a standard set of tests for both Bucket.ls and Bucket.list_versions.