Remove dependency from tqdm
Projects like B2CLI should require it directly. The sdk generally works
fine without that dependency which pulls its own dependencies and it is
a problem for some projects. We'll continue to support classes that use
tqdm.

Any existing installations should keep their dependency in their venv
and new versions of the CLI as well as any other console applications
should depend on tqdm directly, if they choose to use it.